proxy/requirements.txt
requests
fastapi
uvicorn
aioredis
pyyaml